+++ MuckArrows/CreateRecipesPatch.cs
@@ -1,11 +1,11 @@
 ﻿namespace MuckArrows;
 
+using BepInEx;
 using HarmonyLib;
 using System;
 using System.Collections.Generic;
 using System.Linq;
 using UnityEngine;
-using BepInEx;
 
 public class CreateRecipesPatch
 {
@@ -20,27 +20,24 @@ public class CreateRecipesPatch
 		InventoryItem rock = ItemManager.Instance.allItems[90];
 		InventoryItem darkOakWood = ItemManager.Instance.allItems[92];
 		InventoryItem wood = ItemManager.Instance.allItems[94];
+		InventoryItem oakWood = ItemManager.Instance.allItems[95];
 		InventoryItem arrowAdamantite = ItemManager.Instance.allItems[131];
-		InventoryItem arrowFire = ItemManager.Instance.allItems[132];
 		InventoryItem arrowFlint = ItemManager.Instance.allItems[133];
-		InventoryItem arrowLightning = ItemManager.Instance.allItems[134];
 		InventoryItem arrowMithril = ItemManager.Instance.allItems[135];
 		InventoryItem arrowSteel = ItemManager.Instance.allItems[136];
-		InventoryItem arrowWater = ItemManager.Instance.allItems[137];
 
 		arrowFlint.attackDamage = Plugin.FlintArrowDamage;
 		arrowSteel.attackDamage = Plugin.SteelArrowDamage;
 		arrowMithril.attackDamage = Plugin.MithrilArrowDamage;
 		arrowAdamantite.attackDamage = Plugin.AdamantiteArrowDamage;
 
-		// TODO actually make some properly coloured sprites for the arrows. When we've done that, just clone the flint arrow.
 		// All arrows are always added, to avoid any issues with IDs changing if the user were to save, exit, change settings, then re-enter the game
 		List<InventoryItem> newArrows = new()
 		{
-			NewArrow(arrowFlint, "Stone Arrow", "blunt arrows for a long distance slapp", Plugin.StoneArrowDamage, 4, arrowFlint.sprite, Plugin.CanCraftStoneArrows,  new() { amount = 2, item = wood }, new() { amount = 1, item = rock }),
-			NewArrow(arrowFlint, "Obamium Arrow", "the 44th arrow ever invented", Plugin.ObamiumArrowDamage, 4, arrowWater.sprite, Plugin.CanCraftObamiumArrows, new() { amount = 2, item = darkOakWood }, new() { amount = 1, item = obamiumBar } ),
-			NewArrow(arrowFlint, "Ruby Arrow", "shooting your arrows with style", Plugin.RubyArrowDamage, 4, arrowFire.sprite, Plugin.CanCraftRubyArrows, new() { amount = 2, item = darkOakWood }, new() { amount = 1, item = ruby } ),
-			NewArrow(arrowFlint, "Chunkium Arrow", "how do these even fly?", Plugin.ChunkiumArrowDamage, 4, arrowLightning.sprite, Plugin.CanCraftChunkiumArrows, new() { amount = 2, item = darkOakWood }, new() { amount = 1, item = chunkiumBar } ),
+			NewArrow(arrowFlint, "Stone Arrow", "blunt arrows for a long distance slapp", Plugin.StoneArrowDamage, 4, arrowFlint.sprite, arrowFlint.sprite.texture, Plugin.CanCraftStoneArrows,  new() { amount = 2, item = wood }, new() { amount = 1, item = rock }),
+			NewArrow(arrowFlint, "Obamium Arrow", "the 44th arrow ever invented", Plugin.ObamiumArrowDamage, 4, Resources.ObamiumArrowSprite, Resources.ObamiumArrowTexture, Plugin.CanCraftObamiumArrows, new() { amount = 2, item = darkOakWood }, new() { amount = 1, item = obamiumBar } ),
+			NewArrow(arrowFlint, "Ruby Arrow", "shooting your arrows with style", Plugin.RubyArrowDamage, 4, Resources.RubyArrowSprite, Resources.RubyArrowTexture, Plugin.CanCraftRubyArrows, new() { amount = 2, item = darkOakWood }, new() { amount = 1, item = ruby } ),
+			NewArrow(arrowFlint, "Chunkium Arrow", "how do these even fly?", Plugin.ChunkiumArrowDamage, 4, Resources.ChunkiumArrowSprite, Resources.ChunkiumArrowTexture, Plugin.CanCraftChunkiumArrows, new() { amount = 2, item = oakWood }, new() { amount = 1, item = chunkiumBar } ),
 		};
 
 		int id = ItemManager.Instance.allScriptableItems.Length;
@@ -53,7 +50,7 @@ public class CreateRecipesPatch
 		ItemManager.Instance.allScriptableItems = ItemManager.Instance.allScriptableItems.Concat(newArrows).ToArray();
 		ArrowRecipes = newArrows.ToArray();
 	}
-	private static InventoryItem NewArrow(InventoryItem baseArrow, string name, string description, int attackDamage, int amount, Sprite sprite, bool craftable, params InventoryItem.CraftRequirement[] requirements)
+	private static InventoryItem NewArrow(InventoryItem baseArrow, string name, string description, int attackDamage, int amount, Sprite sprite, Texture2D texture, bool craftable, params InventoryItem.CraftRequirement[] requirements)
 	{
 		InventoryItem a = ScriptableObject.CreateInstance<InventoryItem>();
 		a.amount = amount;
@@ -92,7 +89,14 @@ public class CreateRecipesPatch
 		a.rotationOffset = baseArrow.rotationOffset;
 		a.scale = baseArrow.scale;
 		a.sharpness = baseArrow.sharpness;
+
+		a.material = new Material(baseArrow.material)
+		{
+			mainTexture = texture,
+			color = new Color(1f, 1f, 1f, 1f)
+		};
 		a.sprite = sprite;
+
 		a.stackable = baseArrow.stackable;
 		a.stamina = baseArrow.stamina;
 		a.stationRequirement = baseArrow.stationRequirement;