+++ MuckArrows/CHANGELOG.md
@@ -1,2 +1,6 @@
 # v1.0.0
 - Initial Release
+
+# v1.1.0
+- Sprites for Ruby, Obamium, and Chunkium arrows added
+- Chunkium arrows are now made using Oak Wood